Implement the Plausible wrapper
This implements a Plausible service that wraps the plausible-tracker
package. This service makes it easy to use in Ember apps and it provides
some extra features that aren't available in the base package yet.

The default configuration matches the behavior of the standalone Plausible
script to make the transition and required setup
as simple as possible. The plausible-tracker package is lazy-loaded so
it's only downloaded when needed.